+++ internal/analyzer/docker.go
@@ -10,10 +10,9 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/kvesta/vesta/pkg/vulnlib"
-
 	"github.com/docker/docker/api/types"
 	version2 "github.com/hashicorp/go-version"
+	"github.com/kvesta/vesta/pkg/vulnlib"
 	"github.com/tidwall/gjson"
 )